Moved the fetch messages method to message service
